Move Atmosphere envmap light to bevy_light and relax query filters
